Fix page height calcuation for documents with height-restricted body
In case the <body> element of a page has fixed height defined
(e.g.: `height: 100%; overflow: auto`), the package would incorrectly assume
that the document is only one frame height high in the y-axis which would
result in cropping of content. This change, although not very performant,
calculates the true content height based on the largest body child element.